Convert all PME-tables to use AJAX calls for updating
The table itself still needs to be reloaded, but all JS script stuff and
the page header can stay in tact. It also feels better, even at points
where it is not actually faster. Also: add/change/delete/view are now
popups, which also spares a couple of page reloads.