Revise instrument table.
Very much simpler, now that we only use this table as the primary pivot
table for the instruments, no more set enum fields, yeah!

We also allow deletion of unused instruments (e.g. if erroneously adding
a new one or so). Used instruments may not be deleted. Eventually we may
introduce a "disabled" state to mask them out.