Allow omitting actual join for values, update fixes
If values[table] is acutally only needed to define the select values for
the option drop-down, then there is no need to do the actual join. Will
not work for VIEW and LIST mode, as there the description field is used,
which needs the actual join. The application has to take care of this.