PageRenderer: provide a generic pre-commit action queue with undo stack.
Idea is to run action just before committing the current transaction. A
failing action will then trigger a rollback as well as the undo-actions
of all previously run pre-commit actions in reverse order.

If the final transaction commit should fail, then all undo-actions are
also executing.

A common example would be a folder rename after changing e.g. the name
of a project or the userId-slug of a musician.