DB Link the project-payments to the option-keys which define receivables, read on ...
The idea is that there must not be any payment without a defined reason
for the payment. The money to be received must first be defined by a
service-fee option field.

This also implies that "composite debit-notes" for several or all
service-fee options will have to be split (in the data-base, not the
bank-transfer) into their respective parts.

So one debit-note may result in several ProjectPayment entries which
refer many-to-one to a defined service-fee option.

The amount to debit for a specific service-fee option is then just the
difference between the amount already paid for this option and the
amount to pay. Likewise, if too much has been debited, then it is
comparatively easy to compute the amount of the reimbursement needed for
this field.