Cleanup deposits for service fees and cleanup default value handling.
Remov deposit as separate "field-type", instead attach an optional
deposit and date to each service-fee field. In the UI the deposit value
is shown and required if the deposit due-date is configured.

Default values are now always references to a specific
ProjectParticipantFieldDataOption. This is natural as even simple inputs
have at least one such option.

Option and data have one additional field for the deposit value.