Cloud/session closing: optionally render before returning response.
Some of the legacy PME code needs access to the session. However, the
page is generated using callbacks in the "output phase" after the
post-controller middleware already has closed the session. We cope with
this by a "PreRendereredTemplateResponse" which just renders before
returning the response if the session is required.