JS PME Dialogs: try to compensate jQuery dialog width when vertical scrollbar is present.
When a vertical scrollbar is needed because of vertical overflow it may
happen that the horizontal width is not adjusted, even if all ambient
container divs are styled with width=auto. Use Javascript to correct it.

This should be more-or-less ok as jQuery adds a width-style as soon as
the dialog is moved anyway.