Delete musicians: introduce an MusicianService::impersonateMusician() method.
The idea is to be able to remove all personal data while maintaining the
project structure of old projects. "impersonate" means to remove all
file-system data, all contact information, set the musician to "soft
deleted" and use its UUID as name and user id. The email address points
to the orchestra email address with an "+...." addition.