Hopefully complete the link-share-large-attachments stuff by adding clean-up hooks.
Large attachments are link-shared if beyond a configurable size (can
even be disabled). Attachments are saved in the cloud and shared by link
for a configurable time (can even be set to infinity).

Orphan (preview) and expired links are cleaned by a cron job after the
link-share has expired if the file has an age of at least 24 hours.

The latter avoids race-conditions when composing messages from different
web-browsers. We do not want to destroy the preview email of another
user just because one web-interface has triggered a cleanup when closing
the email composition UI.