Bugfix: force directory modification time to mtime of data-base.
Ugly cludge: currently we are not tracking deletion time of objects in
the collections which define the exported "virtual" files. This is a
problems as NC uses the directory modification time to invalidate its
file-cache. To hack around we set the mtime of all directories to the
mtime of the database. Ugly, inefficient, but works.