Crypto: do not persist the encryption context.
Instead: maintain file-ownership of encrypted files in a join table. The
ownership of the bank-accounts already is available in the
SepaBankAccount entity by means of its composite id (Musician entity).

The problem with the encryption context is that it stores user-ids.
Hence, if a user-id would change we would have to adjust all the
encryption contexts as well.