DB ORM, file entities: work around unsupported lazy-loading for one-to-one relations.
Doctrine\ORM still does not support lazy-loading of one-to-one from the
inverse side, so work around by using one-directional one-to-one
mappings from both sides. This works in this case, as the File entity is
the ForeignKey of the FileData entity, so the id of the File entity is
also the join-column for the FileData entity.

Another possibility would have been a fake One-Many/Many-One pair.