Harden single-row access to private data by an access token.
Scheme is: random access token is generated and stored as hash in the
new MusicianRowAccessTokens table. The un-hashed token is stored
encrypted with the public key of the respective musician.

If the musicians logs in it can decrypt the access token and get access
to its data via views which restrict the access to just the current
users data.

In addition to the token also the user-id is checked and must match the
user-id stored in the musicians table.