Participant-fields: add access-restriction in order NOT to export all fields to the member's cloud app.
Now that the member's app is close to becoming functional we need to
make sure that internal data is not leaked to a participant.

This is made sure by an additional field which determines whether a
data-option is shared with the respective person. Shared items shoudl
be:

instrument insurances and related data
maybe some member data like the date of joining the orchestra (and
related documents).