Do not translate names used to form file-system entries.
Clearly not a good idea: the file-system entries are unique by name, so
the field and option names linked to file-system entries should just not
be translated.

Gedmo:
Allow overide of the list of translated fields on a per instance basis.
This is used to not-translate fields in ProjectParticipantField entities
which are used to define file-system names.

ParticipantFields:
Use vanilla DB values for file-system entries. Just do not translate fields which are used to define file-system
entries.