ClassMetadataDecorator: when injecting references update the inverse side if possible.
The use of the decorator was to update as much of the join-column as
possible, in particular if join-columns are shared by different
associations. This is now augmented by updating the inverse side if the
generated proxy entities are already fully loaded and if the
corresponding collections on the inverse side are already initialized.
Read: make the association network consistent when it does not cause
databse traffic. In fact, for uninitialized proxies and lazy collections
the association network will be completed by the ORM when loading the
entities from the database.
	# Bitte geben Sie eine Commit-Beschreibung für Ihre Änderungen ein. Zeilen,