EntityManager: introduce post-commit actions run after a successful outermost commit.
The idea is that some actions should only be run after everything else
works ok, e.g. because they cannot be undone and therefor may not be
suitable for running in the pre-commit run-queue. Or just because it
does not matter but they should be executed anyway, for whatever reason.