DB ORM Events: cleanup life-cycle event hacks, provide access to the decorator.
One problem of the decorator stuff is that the entities in their
life-cycle callbacks have only direct access to the vanilla entity
manager, not the decorated one.

We finally implement a kind-of clean solution by adding a static
registration of the decorator relationships. This eliminates the need to
overload all life-cyle callbacks and greatly simplifies the code.