JS tinyMCE: need to use the initial request token as nonce-value
The nonce is part of the CSP and thus fixed by the CSP headers on
initial page load and thus must be the initial template.

I had this already fixed but forgot about it. This time there is also a
comment in the source code that it must be this way.