BUGFIX: multiple email address support was severely broken ...
a) The musician entity cannot back-link to the email-address entity,
because this generates circular constraints: musician needs email
entity, email entity needs musician. But the referenced email address is
not yet there when persisting new musicians.

b) The legacy "persister" in PMETableViewBase tried to generate multiple
new entities, because other parts of the codes generate new entities as
side-effect.

Further thing was that the selectize widget should accept new email
address "on blur" in order to behave closer to an ordinary text input.

Then there are rumours that old firefox versions may not work with the
JS code. But we will see.