WIP: use Doctrine ORM native PHP enum support.
The drawback is that one stiff would have to define custom types in
order to keep the MySQL enums in the database.

But this means then that using the builtin enum support just complicates
things unneccesserily. Nothing wrong with ORM, it is just fine when one
does not want to use MySQL enums.