EmailForm\RecipientsFilter: fix handling of member status filter.
What happened: when invoking the email-form in project-context there are
several possibilities of the possible recipients "user-base". In
particular it is possible to send email to all persons not participating
in the projet. In this case we want by default to exclude "passive,
soloist, conductor" just like when sending mass emails to "all".

Of course, the user can change this, but this commits changes the
default to a sane state where "email sending is disabled" is honoured.