Group-membership / executive board: try to synchronize by listeners.
The goal of this commit is to make the changes needed in the core
minimal. We still need some changes in the UserManager and the
GroupManager, namely that they use a predictable odering of backends, as
well as a change in the LDAP user backend which allows to actually use
UIDs and GIDs which are already present.

However, changes in the frontend in order to move users between backends
are no longer needed. Instead things is done under the hood by event
listeners (there will of course also be a maintenance command in order
to correct thing after something went wront).

Also: users need no longer be present in the LDAP backend in order to
gain membership to special orchestra groups.