DB legacy phpMyEdit.class.php -- fixed groupby_where.
This feature is used in ProjectPayments, the BulkTransactions and the
BankAcounts table in order to retrieve only a single row in non-list
mode. However, ProjectPayments and BulkTransactions also have a fancy
union layout in order to display to implement split-payments (like in
GnuCash) in display the payment parts in the same table. The previous
implementation then inhibited adding new splits. In order to allow this
the group-by-where conditions needed to be moved into the inner "LEFT
JOIN" statements and remove from the outer WHERE part as that killed the
retrieval of the data from the main table.