DB Legacy pme -- "values from table" join feature.
Allow $fdd['values']{filters,having,description} to reference other
tables. To this aim the pme backand support some kind of "rotation" of
the join table statement which make the table-to-fetch-the-values-from
the main table and pulls in all needed other tables.

This is only supported for "simple" left-join chains where the join
table conditions involved only refer to a single other table (multiple
columns is ok). This should make some complicated sub-query unneeded.