Vue: work around vue-router, url params are not reactive.
Why: although the router triggers change event when the parameters
change, the entire plot is in itself inconsistent: the docs advocate
(with double plus "please do this") to attach component props to url
parameters, but then when the parameters change, there is no update to
the attached component. In the docs, there is only a very brief hint
"use a wrapper" to overcome this.