WIP, sort-of-working calendar event popup.
We need to revise the way we handle the history.

The calendar app only ever pushes to the history. Maybe the better
strategy, though at first glimpse I would have voted to use
router.back(-#STEPS) to return to the base view. Mmmh. But there is no
callback hook for router.push() vs router.replace() so the number of
pushes to the browser history is not exposed. Maybe change that ...