PHP-Toolkit: new PreRenderedTemplateResponse.
Purpose: $response->render() is called outside of the middleware loop
and thus exceptions thrown there are not caught by the middleware. Using
$response->preRender() we force rendering the HTML response inside the
controller such that the surrounding middleware can catch exception
thrown during render.

The downside is that more work is then done inside the middle-ware loop.
An outcome whould be to reimplement the render() method and return a
JSON error response.