ExceptionMiddleware: use the afterController() hook to report exceptions to the frontend.
In order to do this we force render for certain response type (caching
the outcome of render()). Otherwise exceptions during render would end
up in the top-level exception handler resulting in the top-level
exception template being rendered. However, in some cases we have
specialized error reporting facilities which we want to be triggered on
errors which we can handle.