Store post data in the PHP session in order to make page-reload work.
As long as the user operates within the programmtic history stack this
is not needed. However, a page reload will zap and restart the
JavaScript engine. So for this case we provide a graceful reload by
using a hash of the previous post-data  in order to really restore the
current view.