Make history replay work.
- push on top of current position
- replace current history
- push on top of end of current history.

There may still be some loose ends, but the stuff seems to be usable.

TODO: the error stack should be simplified. As the key attached to each
state is just an increasing timestamp, the doubly linked list concept
should be replaced by a sorted array (by key). This would also greatly
simplify the storage of the history key in the database.