DynamicSvgIcon.vue: make ids unique to allow multiple inclusion.
At least FF has a problem with different SVG using the same ids. Well,
this is inconvenient, however, as inline SVGs are part of the DOM and
ids are by definition required to be unique a web browser cannot be
blamed for assuming that this is indeed the case. We use a simple regex
based search and replace on the string representing the SVG.