PROJECT ASSOCIATES -- intermediate "finalization"
This commit finalizes the ProjectAssociates view up to a certain working
intermediate state. "intermediate" as the address book integration still
has to be done.

We use the existing ProjectInstrument entity in order to split the
project participants into "real" participants and "associated"
participants, distinguished by a virtual "not an instrument" instrument
family and an "associated" participation status (should be clean up).

For the end-user the visual changes should jsut be an additional
"business contacts" view which ATM (address book integration is missing)
just adds a secondary participation table for associated (legal)
persons.