PageRenderer: participants may disappear in change dialog.
When changing "deleted" to non-null or flipping the participation status
between 'associated' and something else the change dialog may end up
with an empty data row. This should be explained further to the user.

However this commit as first step makes sure that we do not catch
exceptions in this case.