Legacy mounted vue components, fix memory leak.
When mount vue components into legacy code the components are not
destroyed when parts of the DOM is replaced as this bypasses Vue's own
reactivity. We remember the vue components in jQuery data attributes and
destroy them manually before replacing the code. Currently this affects
the project actions menu and the legacy query log.