InstrumentInsuranceService, avoid floats as this causes roundoff problems.
Problematic are things like 0.XX4999999999999998 which result from
unavoidable round-off errors with floating point arithmetic. However, we
do want to have this correctly rounded to 0.X(X+1) upwards. One solution
is to use exact rational arithmetic which is unproblematic here as we
know that the denominator is at most 1.200.000.000 and all possible
denominators are dividers of this maximum.