RationalNumbers, add short-cuts and assigning operations.
As there is unfortunately no operator overloading with PHP we do want to
avoid overly long method names for arithmetic operations. We keep the
long names of the base class and add shorter aliases and add new
operations with short names.

Assigning operations are not very efficient, but make the consuming code
shorter:

$number1->addEq($number2);

is much more handy as

$number1 = $number1->add($number2);