Add balancing accounts for double book-keeping.
This is somewhat GnuCash centric, although it should be possible to
generalize it for other accounting software.

A close integration with GNC is not possible as the SQL data backend of
GNC is slow to an extent that it is not usable for daily work. Until (if
ever) this is fixed a loose integration with import/export data can be
implemented.