PHP, fix Common\Transliterator.
The ASCII-Transliteror should leave all printable ASCII-characters
alone, unfortunately the Symfony AsciiSlugger does the right things but
then replaces any [^0-9a-zA-Z] by the separator. The AsciiSlugger can
not easily be inherited from as all its members are private (save ctor
and slug(). So patch it to also expose the transliteration as method
transliterate() without replacing non-alphanumeric with the separator.