ParticipantFields, enforce default values for new participants.
This cannot be done by legacy PME. On the initial persist of a
participant it gets all the default values.

TODO: when adding new fields the default values should also be enforced
on existing participants.

What cannot be done: when changing default values of existing fields
there is no way to tell whether data attached to participants just
coincides by chance, so existing data can then not be tweaked.