WIP, EntitySerializer.
The goal is to serialize entities including their association relations
in a space efficient way, i.e. normally sparse, but with the possibility
to reconstruct the "association network" incrementally.

This is meant as bridge to the frontend code which already contains
typings generated by Spatie\TypeScriptExporter.

So this serializer will then generate code with entities conforming to
the typings in order not to have to duplicate the class structure "by
hand".