TS type-traits.d.ts: represent non-negative number as tuples.
This is no more costly than using number types and sequences helpers and
reduces the risk of inifinite recursion. So a non-negative integral
number is now a sequence [0, ..., N-1] of integral number literals. The
actual value is the length of the equence. Zero is represented as the
empty tuple []. We only need this for small integers, the recursion
limit of TS is never hit.