TS entity factory / repository & friends, polish types.
Use the optional "depth" parameter to distinguish between entities which
already fetched associations and entities where accessing an association
yields a promise which fetches missing entities. If the "fetch depth" of
an entity is unknown associations still are only promises, but at any
rate the cache of already fetched entities is used first before querying
the backend.